--- a/chia/consensus/block_body_validation.py
+++ b/chia/consensus/block_body_validation.py
@@ -44,6 +44,15 @@ log = logging.getLogger(__name__)
 #           :
 
 
+@dataclass(frozen=True)
+class ForkAdd:
+    coin: Coin
+    confirmed_height: uint32
+    timestamp: uint64
+    hint: Optional[bytes]
+    is_coinbase: bool
+
+
 @dataclass
 class ForkInfo:
     # defines the last block shared by the fork and the main chain. additions
@@ -58,8 +67,7 @@ class ForkInfo:
     # the header hash of the peak block of this fork
     peak_hash: bytes32
     # The additions include coinbase additions
-    # coin, creation-height, timestamp
-    additions_since_fork: Dict[bytes32, Tuple[Coin, uint32, uint64]] = field(default_factory=dict)
+    additions_since_fork: Dict[bytes32, ForkAdd] = field(default_factory=dict)
     # coin-id
     removals_since_fork: Set[bytes32] = field(default_factory=set)
 
@@ -83,14 +91,14 @@ class ForkInfo:
             timestamp = block.foliage_transaction_block.timestamp
             for spend in npc_result.conds.spends:
                 self.removals_since_fork.add(bytes32(spend.coin_id))
-                for puzzle_hash, amount, _ in spend.create_coin:
+                for puzzle_hash, amount, hint in spend.create_coin:
                     coin = Coin(bytes32(spend.coin_id), bytes32(puzzle_hash), uint64(amount))
-                    self.additions_since_fork[coin.name()] = (coin, height, timestamp)
+                    self.additions_since_fork[coin.name()] = ForkAdd(coin, height, timestamp, hint, False)
         for coin in block.get_included_reward_coins():
             assert block.foliage_transaction_block is not None
             timestamp = block.foliage_transaction_block.timestamp
             assert coin.name() not in self.additions_since_fork
-            self.additions_since_fork[coin.name()] = (coin, block.height, timestamp)
+            self.additions_since_fork[coin.name()] = ForkAdd(coin, block.height, timestamp, None, True)
 
 
 async def validate_block_body(
@@ -409,13 +417,13 @@ async def validate_block_body(
             # Check for spending a coin that does not exist in this fork
             log.error(f"Err.UNKNOWN_UNSPENT: COIN ID: {rem} NPC RESULT: {npc_result}")
             return Err.UNKNOWN_UNSPENT, None
-        new_coin, confirmed_height, confirmed_timestamp = fork_info.additions_since_fork[rem]
+        addition: ForkAdd = fork_info.additions_since_fork[rem]
         new_coin_record: CoinRecord = CoinRecord(
-            new_coin,
-            confirmed_height,
+            addition.coin,
+            addition.confirmed_height,
             uint32(0),
             False,
-            confirmed_timestamp,
+            addition.timestamp,
         )
         removal_coin_records[new_coin_record.name] = new_coin_record