This PR is best reviewed one commit at a time.

The main change is in the last commit.

### Purpose:

In `_reconsider_peak()` we check if the new fork is heavier and if so, decide to switch our peak to this new chain. We also return a `StateChangeSummary` indicating to the caller what (if anything) has changed. e.g. any coin spends that were reorged-out and any new additions and removals (to update the mempool) and hints (to update the hint store).

In order to compute the additions and removals, we fetch all the full blocks and block records (again) along the fork and execute all the block generators (again) (`get_tx_removals_and_additions()`). This is wasteful, and especially risky with expensive block generators.

This patch updates the `ForkInfo` object to keep a little bit more state, in order to be able to produce the `StateChangeSummary` without having to re-run the block generators. This saves looking up `FullBlock`s from the database in addition to the CPU of executing blocks.

### Current Behavior:

When switching to a new fork of the chain, we re-run all the block generators along the fork in order to compute additions and removals, even though we already have most of that information now (in `ForkInfo`).

### New Behavior:

While validating blocks on the fork, store confirmation height of additions and removals as well as whether additions are coinbase rewards or not. This lets us build the `StateChangeSummary` object without re-running block generators, and without fetching `FullBlock`s from the database. We still need to fetch the `BlockRecord`s though.

### Testing Notes:

There's a sync-from-scratch test [here](https://grafana.chiatechlab.com/d/oBgwkDI7z/blockchain-sync-tests?orgId=1&var-ref=reorg-optimization&var-repo=All&var-network=mainnet&var-network=testnet10&var-network=testneta&refresh=1h).

Profiling this on an M1 did not give the speed-up I was expecting. It saves about 9% (time) in `test_long_reorg_nodes` test.

### timing M1

| test | before | after | after / before |
| --- | --- | --- | --- |
| reorg1 | 6.47s | 6.02s | 93% |
| reorg2 | 27.13s | 24.77s | 91.3% |

### timing RPi

| test | before | after | after / before |
| --- | --- | --- | --- |
| reorg1 | 61.84s | 58.27s  | 94.38% |
| reorg2 | 274.26s | 251.79s | 91.80% |

### profile (RPi)

It appears a large part of the time is spent waiting for database I/O (i.e. time spent in `kqueue`). Pay attention to the `kqueue` and `_reconsider_peak` timings of the following profiles:

| test | function | before | after |
| --- | --- | --- | --- |
| reorg 1 | _reconsider_peak | 6.51% | 1.72% |
| reorg 1 | add_block | 14.02% | 9.81% |
| reorg 2 | _reconsider_peak | 9.18% | 3.32% |
| reorg 2 | add_block | 17.31% | 12.26% |

### main

![test-reorg-1](https://github.com/Chia-Network/chia-blockchain/assets/661450/0697b411-58af-4399-8768-c4cf198820be)

![test-reorg-2](https://github.com/Chia-Network/chia-blockchain/assets/661450/4b9c6863-02a3-47f5-9406-842d99f4d127)

### with this patch

![test-reorg-1](https://github.com/Chia-Network/chia-blockchain/assets/661450/e7a1bf2d-598b-4b3a-8045-a42b8d112250)

![test-reorg-2](https://github.com/Chia-Network/chia-blockchain/assets/661450/30385c88-754e-4239-860c-9f0e99415e8a)